added GDP pieChart
